requirements.txt
absl-py==2.1.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
comm==0.2.0
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.3
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.2.0
jax==0.4.37
jaxlib==0.4.36
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mediapipe==0.10.20
ml_dtypes==0.5.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
nest-asyncio==1.5.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==23.2
pandas==2.2.3
parso==0.8.3
pillow==11.0.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==4.25.5
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.98
pytz==2024.2
pywin32==306
pyzmq==25.1.1
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
tzdata==2024.2
wcwidth==0.2.12
Werkzeug==3.1.2
